feat(DynamicComponent): edge-to-edge and raw (without figure wrapper)
